Refactor Database model out of superset/models/core.py
