+++ superset/examples/helpers.py
@@ -22,6 +22,7 @@ from io import BytesIO
 from typing import Set
 from urllib import request
 
+import superset.models.database
 from superset import app, db
 from superset.connectors.connector_registry import ConnectorRegistry
 from superset.models import core as models
@@ -29,7 +30,7 @@ from superset.models import core as models
 BASE_URL = "https://github.com/apache-superset/examples-data/blob/master/"
 
 # Shortcuts
-DB = models.Database
+DB = superset.models.database.Database
 Slice = models.Slice
 Dash = models.Dashboard
