+++ superset/models/core.py
@@ -18,25 +18,19 @@
 """A collection of ORM sqlalchemy models for Superset"""
 import json
 import logging
-import textwrap
-from contextlib import closing
-from copy import copy, deepcopy
+from copy import copy
 from datetime import datetime
-from typing import Any, Callable, Dict, List, Optional, Set, Tuple, Type, TYPE_CHECKING
+from typing import Any, Dict, List, Optional, Set, Type, TYPE_CHECKING
 from urllib import parse
 
-import numpy
-import pandas as pd
 import sqlalchemy as sqla
-import sqlparse
-from flask import escape, g, Markup, request
+from flask import escape, Markup
 from flask_appbuilder import Model
 from flask_appbuilder.models.decorators import renders
 from flask_appbuilder.security.sqla.models import User
 from sqlalchemy import (
     Boolean,
     Column,
-    create_engine,
     DateTime,
     ForeignKey,
     Integer,
@@ -45,24 +39,18 @@ from sqlalchemy import (
     Table,
     Text,
 )
-from sqlalchemy.engine import Dialect, Engine, url
-from sqlalchemy.engine.reflection import Inspector
-from sqlalchemy.engine.url import make_url, URL
 from sqlalchemy.orm import relationship, sessionmaker, subqueryload
 from sqlalchemy.orm.session import make_transient
-from sqlalchemy.pool import NullPool
 from sqlalchemy.schema import UniqueConstraint
-from sqlalchemy.sql import Select
-from sqlalchemy_utils import EncryptedType
 
-from superset import app, db, db_engine_specs, is_feature_enabled, security_manager
+from superset import app, db, is_feature_enabled, security_manager
 from superset.connectors.connector_registry import ConnectorRegistry
-from superset.db_engine_specs.base import TimeGrain
 from superset.legacy import update_time_range
+from superset.models.database import Database
 from superset.models.helpers import AuditMixinNullable, ImportMixin
 from superset.models.tags import ChartUpdater, DashboardUpdater, FavStarUpdater
 from superset.models.user_attributes import UserAttribute
-from superset.utils import cache as cache_util, core as utils
+from superset.utils import core as utils
 from superset.viz import BaseViz, viz_types
 
 if TYPE_CHECKING:
@@ -750,541 +738,6 @@ class Dashboard(  # pylint: disable=too-many-instance-attributes
             cls=utils.DashboardEncoder,
             indent=4,
         )
-
-
-class Database(
-    Model, AuditMixinNullable, ImportMixin
-):  # pylint: disable=too-many-public-methods
-
-    """An ORM object that stores Database related information"""
-
-    __tablename__ = "dbs"
-    type = "table"
-    __table_args__ = (UniqueConstraint("database_name"),)
-
-    id = Column(Integer, primary_key=True)  # pylint: disable=invalid-name
-    verbose_name = Column(String(250), unique=True)
-    # short unique name, used in permissions
-    database_name = Column(String(250), unique=True, nullable=False)
-    sqlalchemy_uri = Column(String(1024), nullable=False)
-    password = Column(EncryptedType(String(1024), config["SECRET_KEY"]))
-    cache_timeout = Column(Integer)
-    select_as_create_table_as = Column(Boolean, default=False)
-    expose_in_sqllab = Column(Boolean, default=True)
-    allow_run_async = Column(Boolean, default=False)
-    allow_csv_upload = Column(Boolean, default=False)
-    allow_ctas = Column(Boolean, default=False)
-    allow_dml = Column(Boolean, default=False)
-    force_ctas_schema = Column(String(250))
-    allow_multi_schema_metadata_fetch = Column(  # pylint: disable=invalid-name
-        Boolean, default=False
-    )
-    extra = Column(
-        Text,
-        default=textwrap.dedent(
-            """\
-    {
-        "metadata_params": {},
-        "engine_params": {},
-        "metadata_cache_timeout": {},
-        "schemas_allowed_for_csv_upload": []
-    }
-    """
-        ),
-    )
-    encrypted_extra = Column(EncryptedType(Text, config["SECRET_KEY"]), nullable=True)
-    perm = Column(String(1000))
-    impersonate_user = Column(Boolean, default=False)
-    export_fields = [
-        "database_name",
-        "sqlalchemy_uri",
-        "cache_timeout",
-        "expose_in_sqllab",
-        "allow_run_async",
-        "allow_ctas",
-        "allow_csv_upload",
-        "extra",
-    ]
-    export_children = ["tables"]
-
-    def __repr__(self):
-        return self.name
-
-    @property
-    def name(self) -> str:
-        return self.verbose_name if self.verbose_name else self.database_name
-
-    @property
-    def allows_subquery(self) -> bool:
-        return self.db_engine_spec.allows_subqueries
-
-    @property
-    def allows_cost_estimate(self) -> bool:
-        extra = self.get_extra()
-
-        database_version = extra.get("version")
-        cost_estimate_enabled: bool = extra.get("cost_estimate_enabled")  # type: ignore
-
-        return (
-            self.db_engine_spec.get_allow_cost_estimate(database_version)
-            and cost_estimate_enabled
-        )
-
-    @property
-    def data(self) -> Dict[str, Any]:
-        return {
-            "id": self.id,
-            "name": self.database_name,
-            "backend": self.backend,
-            "allow_multi_schema_metadata_fetch": self.allow_multi_schema_metadata_fetch,
-            "allows_subquery": self.allows_subquery,
-            "allows_cost_estimate": self.allows_cost_estimate,
-        }
-
-    @property
-    def unique_name(self) -> str:
-        return self.database_name
-
-    @property
-    def url_object(self) -> URL:
-        return make_url(self.sqlalchemy_uri_decrypted)
-
-    @property
-    def backend(self) -> str:
-        sqlalchemy_url = make_url(self.sqlalchemy_uri_decrypted)
-        return sqlalchemy_url.get_backend_name()
-
-    @property
-    def metadata_cache_timeout(self) -> Dict[str, Any]:
-        return self.get_extra().get("metadata_cache_timeout", {})
-
-    @property
-    def schema_cache_enabled(self) -> bool:
-        return "schema_cache_timeout" in self.metadata_cache_timeout
-
-    @property
-    def schema_cache_timeout(self) -> Optional[int]:
-        return self.metadata_cache_timeout.get("schema_cache_timeout")
-
-    @property
-    def table_cache_enabled(self) -> bool:
-        return "table_cache_timeout" in self.metadata_cache_timeout
-
-    @property
-    def table_cache_timeout(self) -> Optional[int]:
-        return self.metadata_cache_timeout.get("table_cache_timeout")
-
-    @property
-    def default_schemas(self) -> List[str]:
-        return self.get_extra().get("default_schemas", [])
-
-    @classmethod
-    def get_password_masked_url_from_uri(cls, uri: str):  # pylint: disable=invalid-name
-        sqlalchemy_url = make_url(uri)
-        return cls.get_password_masked_url(sqlalchemy_url)
-
-    @classmethod
-    def get_password_masked_url(
-        cls, url: URL  # pylint: disable=redefined-outer-name
-    ) -> URL:
-        url_copy = deepcopy(url)
-        if url_copy.password is not None:
-            url_copy.password = PASSWORD_MASK
-        return url_copy
-
-    def set_sqlalchemy_uri(self, uri: str) -> None:
-        conn = sqla.engine.url.make_url(uri.strip())
-        if conn.password != PASSWORD_MASK and not custom_password_store:
-            # do not over-write the password with the password mask
-            self.password = conn.password
-        conn.password = PASSWORD_MASK if conn.password else None
-        self.sqlalchemy_uri = str(conn)  # hides the password
-
-    def get_effective_user(
-        self,
-        url: URL,  # pylint: disable=redefined-outer-name
-        user_name: Optional[str] = None,
-    ) -> Optional[str]:
-        """
-        Get the effective user, especially during impersonation.
-        :param url: SQL Alchemy URL object
-        :param user_name: Default username
-        :return: The effective username
-        """
-        effective_username = None
-        if self.impersonate_user:
-            effective_username = url.username
-            if user_name:
-                effective_username = user_name
-            elif (
-                hasattr(g, "user")
-                and hasattr(g.user, "username")
-                and g.user.username is not None
-            ):
-                effective_username = g.user.username
-        return effective_username
-
-    @utils.memoized(watch=("impersonate_user", "sqlalchemy_uri_decrypted", "extra"))
-    def get_sqla_engine(
-        self,
-        schema: Optional[str] = None,
-        nullpool: bool = True,
-        user_name: Optional[str] = None,
-        source: Optional[int] = None,
-    ) -> Engine:
-        extra = self.get_extra()
-        sqlalchemy_url = make_url(self.sqlalchemy_uri_decrypted)
-        sqlalchemy_url = self.db_engine_spec.adjust_database_uri(sqlalchemy_url, schema)
-        effective_username = self.get_effective_user(sqlalchemy_url, user_name)
-        # If using MySQL or Presto for example, will set url.username
-        # If using Hive, will not do anything yet since that relies on a
-        # configuration parameter instead.
-        self.db_engine_spec.modify_url_for_impersonation(
-            sqlalchemy_url, self.impersonate_user, effective_username
-        )
-
-        masked_url = self.get_password_masked_url(sqlalchemy_url)
-        logging.info("Database.get_sqla_engine(). Masked URL: %s", str(masked_url))
-
-        params = extra.get("engine_params", {})
-        if nullpool:
-            params["poolclass"] = NullPool
-
-        # If using Hive, this will set hive.server2.proxy.user=$effective_username
-        configuration: Dict[str, Any] = {}
-        configuration.update(
-            self.db_engine_spec.get_configuration_for_impersonation(
-                str(sqlalchemy_url), self.impersonate_user, effective_username
-            )
-        )
-        if configuration:
-            d = params.get("connect_args", {})
-            d["configuration"] = configuration
-            params["connect_args"] = d
-
-        params.update(self.get_encrypted_extra())
-
-        if DB_CONNECTION_MUTATOR:
-            sqlalchemy_url, params = DB_CONNECTION_MUTATOR(
-                sqlalchemy_url, params, effective_username, security_manager, source
-            )
-        return create_engine(sqlalchemy_url, **params)
-
-    def get_reserved_words(self) -> Set[str]:
-        return self.get_dialect().preparer.reserved_words
-
-    def get_quoter(self):
-        return self.get_dialect().identifier_preparer.quote
-
-    def get_df(  # pylint: disable=too-many-locals
-        self, sql: str, schema: str, mutator: Optional[Callable] = None
-    ) -> pd.DataFrame:
-        sqls = [str(s).strip(" ;") for s in sqlparse.parse(sql)]
-        source_key = None
-        if request and request.referrer:
-            if "/superset/dashboard/" in request.referrer:
-                source_key = "dashboard"
-            elif "/superset/explore/" in request.referrer:
-                source_key = "chart"
-        engine = self.get_sqla_engine(
-            schema=schema, source=utils.sources[source_key] if source_key else None
-        )
-        username = utils.get_username()
-
-        def needs_conversion(df_series: pd.Series) -> bool:
-            return not df_series.empty and isinstance(df_series[0], (list, dict))
-
-        def _log_query(sql: str) -> None:
-            if log_query:
-                log_query(engine.url, sql, schema, username, __name__, security_manager)
-
-        with closing(engine.raw_connection()) as conn:
-            with closing(conn.cursor()) as cursor:
-                for sql_ in sqls[:-1]:
-                    _log_query(sql_)
-                    self.db_engine_spec.execute(cursor, sql_)
-                    cursor.fetchall()
-
-                _log_query(sqls[-1])
-                self.db_engine_spec.execute(cursor, sqls[-1])
-
-                if cursor.description is not None:
-                    columns = [col_desc[0] for col_desc in cursor.description]
-                else:
-                    columns = []
-
-                df = pd.DataFrame.from_records(
-                    data=list(cursor.fetchall()), columns=columns, coerce_float=True
-                )
-
-                if mutator:
-                    df = mutator(df)
-
-                for k, v in df.dtypes.items():
-                    if v.type == numpy.object_ and needs_conversion(df[k]):
-                        df[k] = df[k].apply(utils.json_dumps_w_dates)
-                return df
-
-    def compile_sqla_query(self, qry: Select, schema: Optional[str] = None) -> str:
-        engine = self.get_sqla_engine(schema=schema)
-
-        sql = str(qry.compile(engine, compile_kwargs={"literal_binds": True}))
-
-        if (
-            engine.dialect.identifier_preparer._double_percents  # pylint: disable=protected-access
-        ):
-            sql = sql.replace("%%", "%")
-
-        return sql
-
-    def select_star(  # pylint: disable=too-many-arguments
-        self,
-        table_name: str,
-        sql: Optional[str] = None,
-        schema: Optional[str] = None,
-        limit: int = 100,
-        show_cols: bool = False,
-        indent: bool = True,
-        latest_partition: bool = False,
-        cols: Optional[List[Dict[str, Any]]] = None,
-    ):
-        """Generates a ``select *`` statement in the proper dialect"""
-        eng = self.get_sqla_engine(
-            schema=schema, source=utils.sources.get("sql_lab", None)
-        )
-        return self.db_engine_spec.select_star(
-            self,
-            table_name,
-            sql=sql,
-            schema=schema,
-            engine=eng,
-            limit=limit,
-            show_cols=show_cols,
-            indent=indent,
-            latest_partition=latest_partition,
-            cols=cols,
-        )
-
-    def apply_limit_to_sql(self, sql: str, limit: int = 1000) -> str:
-        return self.db_engine_spec.apply_limit_to_sql(sql, limit, self)
-
-    def safe_sqlalchemy_uri(self) -> str:
-        return self.sqlalchemy_uri
-
-    @property
-    def inspector(self) -> Inspector:
-        engine = self.get_sqla_engine()
-        return sqla.inspect(engine)
-
-    @cache_util.memoized_func(
-        key=lambda *args, **kwargs: "db:{}:schema:None:table_list",
-        attribute_in_key="id",
-    )
-    def get_all_table_names_in_database(
-        self, cache: bool = False, cache_timeout: bool = None, force=False
-    ) -> List[utils.DatasourceName]:
-        """Parameters need to be passed as keyword arguments."""
-        if not self.allow_multi_schema_metadata_fetch:
-            return []
-        return self.db_engine_spec.get_all_datasource_names(self, "table")
-
-    @cache_util.memoized_func(
-        key=lambda *args, **kwargs: "db:{}:schema:None:view_list",
-        attribute_in_key="id",  # type: ignore
-    )
-    def get_all_view_names_in_database(
-        self, cache: bool = False, cache_timeout: bool = None, force: bool = False
-    ) -> List[utils.DatasourceName]:
-        """Parameters need to be passed as keyword arguments."""
-        if not self.allow_multi_schema_metadata_fetch:
-            return []
-        return self.db_engine_spec.get_all_datasource_names(self, "view")
-
-    @cache_util.memoized_func(
-        key=lambda *args, **kwargs: f"db:{{}}:schema:{kwargs.get('schema')}:table_list",  # type: ignore
-        attribute_in_key="id",
-    )
-    def get_all_table_names_in_schema(
-        self,
-        schema: str,
-        cache: bool = False,
-        cache_timeout: int = None,
-        force: bool = False,
-    ) -> List[utils.DatasourceName]:
-        """Parameters need to be passed as keyword arguments.
-
-        For unused parameters, they are referenced in
-        cache_util.memoized_func decorator.
-
-        :param schema: schema name
-        :param cache: whether cache is enabled for the function
-        :param cache_timeout: timeout in seconds for the cache
-        :param force: whether to force refresh the cache
-        :return: list of tables
-        """
-        try:
-            tables = self.db_engine_spec.get_table_names(
-                database=self, inspector=self.inspector, schema=schema
-            )
-            return [
-                utils.DatasourceName(table=table, schema=schema) for table in tables
-            ]
-        except Exception as e:  # pylint: disable=broad-except
-            logging.exception(e)
-
-    @cache_util.memoized_func(
-        key=lambda *args, **kwargs: f"db:{{}}:schema:{kwargs.get('schema')}:view_list",  # type: ignore
-        attribute_in_key="id",
-    )
-    def get_all_view_names_in_schema(
-        self,
-        schema: str,
-        cache: bool = False,
-        cache_timeout: int = None,
-        force: bool = False,
-    ) -> List[utils.DatasourceName]:
-        """Parameters need to be passed as keyword arguments.
-
-        For unused parameters, they are referenced in
-        cache_util.memoized_func decorator.
-
-        :param schema: schema name
-        :param cache: whether cache is enabled for the function
-        :param cache_timeout: timeout in seconds for the cache
-        :param force: whether to force refresh the cache
-        :return: list of views
-        """
-        try:
-            views = self.db_engine_spec.get_view_names(
-                database=self, inspector=self.inspector, schema=schema
-            )
-            return [utils.DatasourceName(table=view, schema=schema) for view in views]
-        except Exception as e:  # pylint: disable=broad-except
-            logging.exception(e)
-
-    @cache_util.memoized_func(
-        key=lambda *args, **kwargs: "db:{}:schema_list", attribute_in_key="id"
-    )
-    def get_all_schema_names(
-        self, cache: bool = False, cache_timeout: int = None, force: bool = False
-    ) -> List[str]:
-        """Parameters need to be passed as keyword arguments.
-
-        For unused parameters, they are referenced in
-        cache_util.memoized_func decorator.
-
-        :param cache: whether cache is enabled for the function
-        :param cache_timeout: timeout in seconds for the cache
-        :param force: whether to force refresh the cache
-        :return: schema list
-        """
-        return self.db_engine_spec.get_schema_names(self.inspector)
-
-    @property
-    def db_engine_spec(self) -> Type[db_engine_specs.BaseEngineSpec]:
-        return db_engine_specs.engines.get(self.backend, db_engine_specs.BaseEngineSpec)
-
-    @classmethod
-    def get_db_engine_spec_for_backend(
-        cls, backend
-    ) -> Type[db_engine_specs.BaseEngineSpec]:
-        return db_engine_specs.engines.get(backend, db_engine_specs.BaseEngineSpec)
-
-    def grains(self) -> Tuple[TimeGrain, ...]:
-        """Defines time granularity database-specific expressions.
-
-        The idea here is to make it easy for users to change the time grain
-        from a datetime (maybe the source grain is arbitrary timestamps, daily
-        or 5 minutes increments) to another, "truncated" datetime. Since
-        each database has slightly different but similar datetime functions,
-        this allows a mapping between database engines and actual functions.
-        """
-        return self.db_engine_spec.get_time_grains()
-
-    def get_extra(self) -> Dict[str, Any]:
-        extra: Dict[str, Any] = {}
-        if self.extra:
-            try:
-                extra = json.loads(self.extra)
-            except json.JSONDecodeError as e:
-                logging.error(e)
-                raise e
-        return extra
-
-    def get_encrypted_extra(self):
-        encrypted_extra = {}
-        if self.encrypted_extra:
-            try:
-                encrypted_extra = json.loads(self.encrypted_extra)
-            except json.JSONDecodeError as e:
-                logging.error(e)
-                raise e
-        return encrypted_extra
-
-    def get_table(self, table_name: str, schema: Optional[str] = None) -> Table:
-        extra = self.get_extra()
-        meta = MetaData(**extra.get("metadata_params", {}))
-        return Table(
-            table_name,
-            meta,
-            schema=schema or None,
-            autoload=True,
-            autoload_with=self.get_sqla_engine(),
-        )
-
-    def get_columns(
-        self, table_name: str, schema: Optional[str] = None
-    ) -> List[Dict[str, Any]]:
-        return self.db_engine_spec.get_columns(self.inspector, table_name, schema)
-
-    def get_indexes(
-        self, table_name: str, schema: Optional[str] = None
-    ) -> List[Dict[str, Any]]:
-        return self.inspector.get_indexes(table_name, schema)
-
-    def get_pk_constraint(
-        self, table_name: str, schema: Optional[str] = None
-    ) -> Dict[str, Any]:
-        return self.inspector.get_pk_constraint(table_name, schema)
-
-    def get_foreign_keys(
-        self, table_name: str, schema: Optional[str] = None
-    ) -> List[Dict[str, Any]]:
-        return self.inspector.get_foreign_keys(table_name, schema)
-
-    def get_schema_access_for_csv_upload(  # pylint: disable=invalid-name
-        self
-    ) -> List[str]:
-        return self.get_extra().get("schemas_allowed_for_csv_upload", [])
-
-    @property
-    def sqlalchemy_uri_decrypted(self) -> str:
-        conn = sqla.engine.url.make_url(self.sqlalchemy_uri)
-        if custom_password_store:
-            conn.password = custom_password_store(conn)
-        else:
-            conn.password = self.password
-        return str(conn)
-
-    @property
-    def sql_url(self) -> str:
-        return f"/superset/sql/{self.id}/"
-
-    def get_perm(self) -> str:
-        return f"[{self.database_name}].(id:{self.id})"
-
-    def has_table(self, table: Table) -> bool:
-        engine = self.get_sqla_engine()
-        return engine.has_table(table.table_name, table.schema or None)
-
-    def has_table_by_name(self, table_name: str, schema: Optional[str] = None) -> bool:
-        engine = self.get_sqla_engine()
-        return engine.has_table(table_name, schema)
-
-    @utils.memoized
-    def get_dialect(self) -> Dialect:
-        sqla_url = url.make_url(self.sqlalchemy_uri_decrypted)
-        return sqla_url.get_dialect()()
 
 
 sqla.event.listen(Database, "after_insert", security_manager.set_perm)
