+++ tests/model_tests.py
@@ -21,7 +21,7 @@ import pandas
 from sqlalchemy.engine.url import make_url
 
 from superset import app
-from superset.models.core import Database
+from superset.models.database import Database
 from superset.utils.core import get_example_database, QueryStatus
 
 from .base_tests import SupersetTestCase
