+++ tests/security_tests.py
@@ -24,7 +24,8 @@ from superset import app, appbuilder, db, security_manager, viz
 from superset.connectors.druid.models import DruidCluster, DruidDatasource
 from superset.connectors.sqla.models import SqlaTable
 from superset.exceptions import SupersetSecurityException
-from superset.models.core import Database, Slice
+from superset.models.core import Slice
+from superset.models.database import Database
 from superset.utils.core import get_example_database
 
 from .base_tests import SupersetTestCase
